pyproject.toml
[tool.poetry]
name = "crm-api"
version = "0.1.0"
description = "A test project featuring a CRM API."
authors = ["Philemon Katsarkas <philemon@katsarkas.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
django = "^5.0.7"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"
# django-allauth = {extras = ["socialaccount"], version = "^64.0.0"}
django-allauth = {extras = ["socialaccount"], version = "^0.61.1"}
dj-rest-auth = "^6.0.0"
django-extensions = "^3.2.3"
pillow = "^10.4.0"
gunicorn = "^22.0.0"
django-cleanup = "^8.1.0"
drf-yasg = "^1.21.7"
pre-commit = "^3.8.0"
ruff = "^0.5.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
